refactor: layout 컴포넌트를 없애고 outlet 적용
