+++ src/router/AppRouter.tsx
@@ -1,12 +1,18 @@
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import App from '../App';
 import { PAGE_INFORMATION } from '../constants/page';
+import ProductPage from '../components/page/ProductPage';
 
 const router = createBrowserRouter([
   {
-    path: PAGE_INFORMATION.main.path,
+    path: '',
     element: <App />,
-    children: [],
+    children: [
+      {
+        path: PAGE_INFORMATION.main.path,
+        element: <ProductPage />,
+      },
+    ],
   },
 ]);
